pyproject.toml
[tool.poetry]
name = "todo-project"
version = "0.1.0"
description = "Com esse projeto temos a inteção de estudar a criação de API's com fastAPI."
authors = ["Vicente Marçal <vicente.marcal@gmail.com>"]
license = "GPLv3+"
readme = "README.md"
packages = [{include = "todo"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
sqlmodel = "^0.0.8"
dynaconf = "^3.2.0"
alembic = "^1.11.1"
typer = "^0.9.0"
rich = "^13.5.2"
python-slugify = "^8.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
ipython = "^8.14.0"
ipdb = "^0.13.13"
flake8 = "^6.0.0"
pytest-order = "^1.1.0"
httpx = "^0.24.1"
pytest-cov = "^4.1.0"
taskipy = "^1.11.0"
flake8-pyproject = "^1.2.3"

[tool.poetry.scripts] 
todo = "todo.cli:main"

[tool.pytest.ini_options]
python_files = [
    "test*.py",
    "*tests.py",
]
minversion = "6.0"
addopts = "-ra -q -vv"
testpaths = [
    "/tests",
]

[tool.taskipy.variables]
pyc = "./ -name '*.pyc' -delete"
pycache = "./ -name '__pycache__' -delete"
thumbs = "./ -name 'Thumbs.db' -delete"
root = "./ -name '*~' -delete"

[tool.taskipy.tasks]
test = {cmd = "pytest -x", help = "Run test and abort if has one fail."}
test-cov = {cmd = "pytest --cov=todo", help = "Run test and give us a coverage report."}
lint = {cmd = "flake8 .", help = "Run lint to check PEP8."}
clean = {cmd = "find {pyc} && find {pycache} && find {thumbs} && find {root}", help = "Clear the project off all files that are dispensables.", use_vars = true}
post_clean = {cmd = "rm -rf .cache && rm -rf .pytest_cache && rm -rf dist && rm -rf *.egg-info && rm -rf htmlcov && rm -rf .tox/ && rm -rf site", help = "Remove all files that are dispensables."}
down = {cmd = "docker-compose down", help = "Down the services containers."}
remove-img = {cmd = "docker image rm todo_project_api:latest todo_postgres-15-alpine-multi-user:latest", help = "Delete containers images to build again."}
up = {cmd = "docker-compose up -d", help = "Up the services containers."}
pre_build = "task down && task remove-img"
build = {cmd = "docker-compose build --no-cache", help = "Build the images' services."}
post_build = "task up"

[tool.flake8]
max-line-length = 120
exclude = '.venv/, migrations, .git, __pycache__,'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
